+++ game/ts/api.ts
@@ -21,9 +21,7 @@ import { InternalStressTestAPI } from "../../shared/api/internal/internal-stress
 import { DebugAllColorsAPI } from "../../shared/api/internal/internal-all-colors";
 import { BlankExampleAPI } from "../../shared/api/blank";
 import { disposeServerConfigGui, reRenderServerConfigGui } from "./settings-server-config";
-import { Store } from "../../shared/store";
 import { ChunkedStore } from "../../shared/store-chunk";
-import { startSearch } from "./element-game/search";
 import { startRandomSuggestions } from "./element-game/randomSuggestions";
 
 // @ts-ignore
@@ -199,9 +197,6 @@ export async function connectApi(baseUrl: string, config: ElementalConfig, ui?: 
     await InitElementNews();
 
     await onSaveFileLoad(ui);
-
-    ui.status("Starting Search");
-    await startSearch();
 
     ui.status("Starting Random Suggestions");
     await startRandomSuggestions();